+++ doc/whats_new.rst
@@ -30,4 +30,7 @@ Bug
 API
 ~~~
 
+- Added `ch_types` argument to :func:`autoreject.get_rejection_threshold` to find
+  rejection thresholds for only subset of channel types in the data, by `Mainak Jas` in `#140 <https://github.com/autoreject/autoreject/pull/140>`_
+
 .. _Mainak Jas: https://perso.telecom-paristech.fr/mjas/